Socket boilerplate for Netlink
This commit adds Netlink to SocketRef and SocketRefMut, adds all the
boilerplate methods for NetlinkSocket, and adds the protocol states to
NetlinkSocket.

There are two protocol states in NetlinkSocket: Initial and Closed,
which is similar to the stateless Unix domain socket because, as
specified in netlink(7), Netlink is a datagram-oriented service.