Implement Netlink bind syscall
According to netlink(7), there are two issues worth mentioning.

If the pid of the address is zero, the kernel takes care of assigning it, but we will
leave it untouched at the moment. We can implement the assignment later when we want
to support it.

If the groups of the address is non-zero, it means that the socket wants to listen to
some groups. Since we don't support broadcasting to groups yet, we will emit the error
intead.