Handle RTM_GETLINK
Previously we handle only RTM_GETADDR, this commit now enables us to
handle RTM_GETLINK.

Firstly, we refactored the RTM_GETADDR handling code in recvmsg into a
new method called handle_ifaddrmsg. Then, we implement the RTM_GETLINK
handling part in a method called handle_ifinfomsg.

In handle_ifinfomsg, it works as follows:
- Check that the message type is RTM_GETLINK because the only currently
  supported one is RTM_GETLINK
- Check that ifi_family of the message is either AF_UNSPEC or AF_INET
  because we support only IPv4 addresses
- Check that the rest of the fields (except ifi_change) are zeroes
  because they are unsupported
- Send the message containing all the interfaces with the following
  attributes: IFLA_IFNAME, IFLA_TXQLEN, and IFLA_MTU

We can also send the attributes IFLA_ADDRESS and IFLA_BROADCAST for the
MAC address, but we mark it as a todo instead.

The ifi_flags of the loopback interface is IFF_UP, IFF_LOOPBACK, and
IFF_RUNNING.

The ifi_flags of the ethernet interface is IFF_UP, IFF_BROADCAST,
IFF_RUNNING, and IFF_MULTICAST.

All the flags listed above are from what I observe in the `ip addr`
command.