Send attestations to subnets
In every epoch, each validator will send an attestation to the network
attesting the latest block received. Since the number of validators can
be large in the real network, Ethereum was designed in a way that the
attestations are sent to the subnets instead, each corresponding to a
gossipsub topic and a committee.